data loading almost work

+++ dense_correspondence/correspondence_tools/correspondence_finder.py
@@ -495,7 +495,7 @@ def batch_find_pixel_correspondences(img_a_depth, img_a_pose, img_b_depth, img_b
     img_a_depth_torch = torch.squeeze(img_a_depth_torch, 0)
     img_a_depth_torch = img_a_depth_torch.view(-1,1)
 
-
+    uv_a_vec_flattened = uv_a_vec_flattened.type(torch.int64)
     depth_vec = torch.index_select(img_a_depth_torch, 0, uv_a_vec_flattened)*1.0/DEPTH_IM_SCALE
     depth_vec = depth_vec.squeeze(1)
